requirements.txt
fastapi
uvicorn
opencv-python
numpy
scikit-learn
matplotlib
pillow
torch
torchvision
timm
python-multipart
